Add event_new_eventarg().
event_new_eventarg() is like event_new() but passes the event itself as the
argument to the event callback.

It is not possible to use event_new() while passing the event itself as the
argument to the event callback. A workaround is to use event_assign(),
however, this makes code depend on the size of the event structure.